Validation added

+++ app/Http/Controllers/Backend/ShopController.php
@@ -2,19 +2,19 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\Helpers\APIHelper;
-// use App\Http\Requests\CMS\RiddleCreateRequest;
-// use App\Http\Requests\CMS\RiddleUpdateRequest;
 use App\Shop;
+use App\Helpers\APIHelper;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\BackendRequest\CreateShopRequest;
 
 class ShopController extends Controller
 {
     public function index()
     {
         $shops = Shop::paginate(10);
-        return view('backend.pages.shops.index',["shops"=>$shops]);
+        return view('backend.pages.shops.index', ["shops" => $shops]);
     }
 
     /**
@@ -24,7 +24,12 @@ class ShopController extends Controller
      */
     public function create()
     {
-        return view('backend.pages.shops.create');
+        if (Auth::user()->user_type = 'super_admin') {
+
+            return view('backend.pages.shops.create');
+        } else {
+            return view('backend.pages.shops.index')->with(session()->flash('error', 'You are not authorized to create a shop'));
+        }
     }
 
     /**
@@ -33,7 +38,7 @@ class ShopController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CreateShopRequest $request)
     {
         $shop = new Shop();
         $shop->name = $request->input("name");
@@ -67,8 +72,12 @@ class ShopController extends Controller
      */
     public function edit($id)
     {
-        $shop = Shop::find($id);
-        return view('backend.pages.shops.edit',["shop"=>$shop]);
+        if (Auth::user()->user_type = 'super_admin') {
+            $shop = Shop::find($id);
+            return view('backend.pages.shops.edit', ["shop" => $shop]);
+        } else {
+            return view('backend.pages.shops.index')->with(session()->flash('error', 'You are not authorized to edit a shop'));
+        }
     }
 
     /**
@@ -78,7 +87,7 @@ class ShopController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(CreateShopRequest $request, $id)
     {
         $shop = Shop::find($id);
         $shop->name = $request->input("name");
@@ -101,11 +110,15 @@ class ShopController extends Controller
      */
     public function destroy($id)
     {
-        $result = Shop::find($id)->delete();
-        if ($result) {
-            return redirect()->route('backend.shops.index')->with(session()->flash('success', 'Shop Deleted!'));
+        if (Auth::user()->user_type = 'super_admin') {
+            $result = Shop::find($id)->delete();
+            if ($result) {
+                return redirect()->route('backend.shops.index')->with(session()->flash('success', 'Shop Deleted!'));
+            } else {
+                return redirect()->route('backend.shops.index')->with(session()->flash('error', 'Something went wrong!'));
+            };
         } else {
-            return redirect()->route('backend.shops.index')->with(session()->flash('error', 'Something went wrong!'));
+            return view('backend.pages.shops.index')->with(session()->flash('error', 'You are not authorized to edit a shop'));
         }
     }
 }
